+++ include/rocksdb/table.h
@@ -365,6 +365,11 @@ struct BlockBasedTableOptions {
   // This must generally be true for gets to be efficient.
   bool whole_key_filtering = true;
 
+  // If true, detect corruption during (new) Bloom Filter and
+  // Ribbon Filter's construction by adding checksum on hash entries
+  // of the filter as well as re-querying the filter later.
+  bool detect_filter_construct_corruption = false;
+
   // Verify that decompressing the compressed block gives back the input. This
   // is a verification mode that we use to detect bugs in compression
   // algorithms.